OhdsiRTools-package {OhdsiRTools}	R Documentation

OhdsiRTools: Tools Used by Observational Health Data Science and Informatics (OHDSI)

Description

Includes functions to format and check syntax of R code and packages following the 'OHDSI' R style guidelines. Generate renv lock files for OHDSI study packages.

Author(s)

Maintainer: Martijn Schuemie schuemie@ohdsi.org

Authors:

Marc Suchard

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

http://ohdsi.github.io/OhdsiRTools/

https://github.com/OHDSI/OhdsiRTools

Report bugs at https://github.com/OHDSI/OhdsiRTools/issues

[Package OhdsiRTools version 2.0.1 Index]