Check all code in a package

Check all code in a package

checkUsagePackage(
  package,
  ignoreHiddenFunctions = TRUE,
  suppressBindingKeywords = c("ggplot2", "ffwhich", "subset.ffdf", "glm")
)

Arguments

package
The name of the package to check.

ignoreHiddenFunctions
Ignore functions for which the definition cannot be retrieved?

suppressBindingKeywords
A set of keywords that are indicative of non-standard evaluation.

Details

This function uses the codetools package to check the code from problems. Heuristics are used to eliminate false positives due to non-standard evaluation.

Examples

checkUsagePackage("OhdsiRTools")

#> Warning: Ignoring problem in hidden function '.checkBaseUrl : <anonymous>: local variable 'result' assigned but may not be used'

#> Warning: Ignoring problem in hidden function 'getCohortGenerationStatuses : <anonymous>: local variable 'status' assigned but may not be used'

#> No problems found