createConceptSetWorkbook {OhdsiRTools}	R Documentation

Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook

Description

Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook

Usage

createConceptSetWorkbook(
  conceptSetIds,
  workFolder = NULL,
  baseUrl,
  included = FALSE,
  mapped = FALSE
)

Arguments

conceptSetIds
A vector of concept set IDs.

workFolder
Directory location where the workbook will be saved, defaults to working directory.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

included
Should included concepts be included in the workbook?

mapped
Should mapped concepts be included in the workbook?

Value

A xlsx workbook (conceptSetExpressions.xlsx) that includes a list of all concept set IDs and names and a worksheet for the concepts in each set. Options to include an included concepts and mapped concepts worksheet for each concept set are available.

[Package OhdsiRTools version 1.10.0 Index]