findNonAsciiStringsInFolder {OhdsiRTools}	R Documentation

Find non-ASCII strings in files

Description

Find non-ASCII string in files

Usage

findNonAsciiStringsInFolder(path = ".", recursive = TRUE, pattern = "*.R$")

Arguments

path
Path to the folder containing the files matching the pattern parameter.

recursive
If TRUE, subfolders will also be searched for files matching the pattern parameter.

pattern
The regular expression to use for selecting files. The default is .R files.

Value

A table listing the lines per file containing non-ASCII characters.

[Package OhdsiRTools version 2.0.1 Index]