formatRText {OhdsiRTools}	R Documentation

Format R code

Description

Format R code

Usage

formatRText(text, width.cutoff = 100)

Arguments

text
A character vector with the R code to be formatted.

width.cutoff
Number of characters that each line should be limited to.

Details

DEPRECRATED. Please use styler::style_text instead.

Value

A character vector with formatted R code.

[Package OhdsiRTools version 2.0.1 Index]