getCohortGenerationStatuses {OhdsiRTools}	R Documentation

Get Cohort Generation Statuses

Description

Get Cohort Generation Statuses

Usage

getCohortGenerationStatuses(baseUrl, definitionIds, sourceKeys)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

definitionIds
A list of cohort definition Ids

sourceKeys
A list of CDM source keys. These can be found in Atlas -> Configure.

Details

Obtains cohort generation statuses for a collection of cohort definition Ids and CDM sources. Useful if running multiple cohort generation jobs that are long-running.

Value

A data frame of cohort generation statuses, start times, and execution durations per definition id and source key.

[Package OhdsiRTools version 1.10.0 Index]