getConceptSetsAndConceptsFromCohort {OhdsiRTools}	R Documentation

Get a list of concept sets and concepts from a cohort definition

Description

Get a list of concept sets and concepts from a cohort definition

Usage

getConceptSetsAndConceptsFromCohort(
  baseUrl,
  definitionId,
  vocabSourceKey = NULL
)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

definitionId
The cohort id to fetch concept sets and concepts from

vocabSourceKey
A mysterious parameter.

Details

For a given cohort definition id, get all concept sets and resolve all concepts from each

Value

A list of concept sets, set names, and concepts

Examples

## Not run: 
# This will obtain a list of concept sets and concepts from a cohort id:

getConceptsFromCohortId(baseUrl = "http://server.org:80/WebAPI", definitionId = 123)

## End(Not run)

[Package OhdsiRTools version 1.10.0 Index]