insertCohortDefinitionInPackage {OhdsiRTools}	R Documentation

Load a cohort definition and insert it into this package

Description

Load a cohort definition and insert it into this package

Usage

insertCohortDefinitionInPackage(
  definitionId,
  name = NULL,
  baseUrl,
  generateStats = FALSE
)

Arguments

definitionId
The number indicating which cohort definition to fetch.

name
The name that will be used for the json and SQL files. If not provided, the name in cohort will be used, but this may not lead to valid file names.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

generateStats
Should the SQL include the code for generating inclusion rule statistics? Note that if TRUE, several additional tables are expected to exists as described in the details.

Details

Load a cohort definition from a WebApi instance and insert it into this package. This will fetch the json object and store it in the 'inst/cohorts' folder, and fetch the template SQL and store it in the 'inst/sql/sql_server' folder. Both folders will be created if they don't exist. When using generateStats = TRUE, the following tables are required to exist when executing the SQL: cohort_inclusion, cohort_inclusion_result, cohort_inclusion_stats, and cohort_summary_stats. Also note that the cohort_inclusion table should be populated with the names of the rules prior to executing the cohort definition SQL.

Examples

## Not run: 
# This will create 'inst/cohorts/Angioedema.json' and 'inst/sql/sql_server/Angioedema.sql':

insertCohortDefinitionInPackage(definitionId = 282,
                                name = "Angioedema",
                                baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package OhdsiRTools version 1.10.0 Index]