invokeCohortSetGeneration {OhdsiRTools}	R Documentation

Invoke the generation of a set of cohort definitions

Description

Invoke the generation of a set of cohort definitions

Usage

invokeCohortSetGeneration(baseUrl, sourceKeys, definitionIds)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

sourceKeys
A list of CDM source keys. These can be found in Atlas -> Configure.

definitionIds
A list of cohort definition Ids

Details

Invokes the generation of a set of cohort definitions across a set of CDMs set up in WebAPI. Use getCohortGenerationStatuses to check the progress of the set.

[Package OhdsiRTools version 1.10.0 Index]