restoreEnvironmentFromPackageOnGithub {OhdsiRTools}	R Documentation

Restore environment stored in package

Description

Restore environment stored in package

Usage

restoreEnvironmentFromPackageOnGithub(
  githubPath,
  pathToCsv = "inst/settings/rEnvironmentSnapshot.csv",
  stopOnWrongRVersion = FALSE,
  strict = FALSE,
  skipLast = TRUE
)

Arguments

githubPath
The path for the GitHub repo containing the package (e.g. 'OHDSI/StudyProtocols/AlendronateVsRaloxifene').

pathToCsv
The path for the snapshot inside the package.

stopOnWrongRVersion
Should the function stop when the wrong version of R is installed? Else just a warning will be thrown when the version doesn't match.

strict
If TRUE, the exact version of each package will installed. If FALSE, a package will only be installed if (a) a newer version is required than currently installed, or (b) the major version number is different.

skipLast
Skip last entry in snapshot? This is usually the study package that needs to be installed manually.

Details

This function restores all packages (and package versions) described in the environment snapshot stored in the package currently being developed. The default location is inst/settings/rEnvironmentSnapshot.csv.

Examples

## Not run: 
restoreEnvironmentFromPackageOnGithub("OHDSI/StudyProtocols/AlendronateVsRaloxifene")

## End(Not run)

[Package OhdsiRTools version 2.0.1 Index]