Take a snapshot of the R environment

Take a snapshot of the R environment

takeEnvironmentSnapshot(rootPackage)

Arguments

rootPackage
The name of the root package

Value

A data frame listing all the dependencies of the root package and their version numbers, in the order in which they should be installed.

Details

This function records all versions used in the R environment that are used by one root package. This can be used for example to restore the environment to the state it was when a particular study package was run using the restoreEnvironment function.

Examples

snapshot <- takeEnvironmentSnapshot("OhdsiRTools")
snapshot

#>        package  version
#> 1            R    3.6.1
#> 2    grDevices    3.6.1
#> 3     graphics    3.6.1
#> 4        utils    3.6.1
#> 5        stats    3.6.1
#> 6      methods    3.6.1
#> 7        rlang    0.4.5
#> 8        tools    3.6.1
#> 9   assertthat    0.2.1
#> 10   backports    1.1.6
#> 11      crayon    1.3.4
#> 12      digest   0.6.25
#> 13    ellipsis    0.3.0
#> 14       fansi    0.4.1
#> 15        glue    1.4.0
#> 16          ps    1.3.2
#> 17          R6    2.4.1
#> 18         sys      3.3
#> 19     askpass      1.1
#> 20         cli    2.0.2
#> 21    processx    3.4.2
#> 22   rprojroot    1.3-2
#> 23        utf8    1.1.4
#> 24       vctrs    0.2.4
#> 25       callr    3.4.3
#> 26        curl      4.3
#> 27        desc    1.2.0
#> 28    jsonlite    1.6.1
#> 29   lifecycle    0.2.0
#> 30    magrittr      1.5
#> 31        mime      0.9
#> 32     openssl    1.4.1
#> 33      pillar    1.4.3
#> 34   pkgconfig    2.0.3
#> 35 prettyunits    1.1.1
#> 36        Rcpp  1.0.4.6
#> 37     stringi    1.4.6
#> 38       withr    2.1.2
#> 39        xfun     0.13
#> 40    evaluate     0.14
#> 41       highr      0.8
#> 42   htmltools    0.4.0
#> 43        httr    1.4.1
#> 44         ini    0.3.1
#> 45       later    1.0.0
#> 46    lazyeval    0.2.2
#> 47    markdown      1.1
#> 48    pkgbuild    1.0.6
#> 49 R.methodsS3    1.8.0
#> 50  rstudioapi     0.11
#> 51     stringr    1.4.0
#> 52      tibble    3.0.0
#> 53        yaml    2.2.1
#> 54        brew    1.0-6
#> 55       clipr    0.7.0
#> 56  commonmark      1.7
#> 57   crosstalk  1.1.0.1
#> 58          fs    1.4.1
#> 59          gh    1.1.0
#> 60       git2r   0.26.1
#> 61 htmlwidgets    1.5.1
#> 62       knitr     1.28
#> 63     pkgload    1.0.2
#> 64      praise    1.0.0
#> 65    promises    1.1.0
#> 66       purrr    0.3.3
#> 67        R.oo   1.23.0
#> 68    rematch2    2.1.1
#> 69         rex    1.1.2
#> 70 sessioninfo    1.1.1
#> 71     whisker      0.4
#> 72        xml2    1.3.1
#> 73       xopen    1.0.0
#> 74        covr    3.5.0
#> 75          DT     0.13
#> 76     memoise    1.1.0
#> 77     R.utils    2.9.2
#> 78   rcmdcheck    1.3.3
#> 79     remotes    2.1.1
#> 80       rJava   0.9-12
#> 81    roxygen2    7.1.0
#> 82   rversions    2.0.1
#> 83    testthat    2.3.2
#> 84     usethis    1.6.0
#> 85         zip    2.0.4
#> 86   codetools   0.2-16
#> 87    devtools    2.3.0
#> 88     formatR      1.7
#> 89       mailR    0.4.1
#> 90    openxlsx    4.1.4
#> 91     RJSONIO  1.3-1.4
#> 92        snow    0.4-3
#> 93         XML 3.99-0.3
#> 94 OhdsiRTools    1.8.0