updateCopyrightYearFile {OhdsiRTools}	R Documentation

Update the copyright year in a R or SQL file

Description

Update the copyright year in a R or SQL file

Usage

updateCopyrightYearFile(file)

Arguments

file
The path to the file.

[Package OhdsiRTools version 2.0.1 Index]