checkUsagePackage {OhdsiRTools}	R Documentation

Check all code in a package

Description

Check all code in a package

Usage

checkUsagePackage(
  package,
  ignoreHiddenFunctions = TRUE,
  suppressBindingKeywords = c("ggplot2", "ffwhich", "subset.ffdf", "glm")
)

Arguments

package
The name of the package to check.

ignoreHiddenFunctions
Ignore functions for which the definition cannot be retrieved?

suppressBindingKeywords
A set of keywords that are indicative of non-standard evaluation.

Details

This function uses the codetools package to check the code from problems. Heuristics are used to eliminate false positives due to non-standard evaluation.

Examples

checkUsagePackage("OhdsiRTools")

[Package OhdsiRTools version 1.10.0 Index]