createRenvLockFile {OhdsiRTools}	R Documentation

Create a renv lock file

Description

Create a renv lock file

Usage

createRenvLockFile(
  rootPackage,
  includeRootPackage = TRUE,
  additionalRequiredPackages = NULL,
  ohdsiGitHubPackages = getOhdsiGitHubPackages(),
  ohdsiStudiesGitHubPackages = rootPackage,
  fileName = "renv.lock"
)

Arguments

rootPackage
The name of the root package, the package that we'd like to be able to run in the end.

includeRootPackage
Include the root package in the renv file?

additionalRequiredPackages
Additional packages we want to have installed (with their dependencies), such as 'keyring'.

ohdsiGitHubPackages
Names of R packages that need to be installed from the OHDSI GitHub.

ohdsiStudiesGitHubPackages
Names of R packages that need to be installed from the OHDSI-Studies GitHub.

fileName
Name of the lock file to be generated.

Details

Create a lock file that allows reconstruction of the R environment using the renv package. This function will include the root file and all of its dependencies in the lock file, requiring the same package versions as currently installed on this computer.

Value

Does not return a value. Is executed for the side-effect of creating the lock file.

[Package OhdsiRTools version 1.10.0 Index]