findNonAsciiStringsInFolder {OhdsiRTools}	R Documentation

Find non-ASCII strings in R files

Description

Find non-ASCII string in R files.

Usage

findNonAsciiStringsInFolder(path = ".", recursive = TRUE)

Arguments

path
Path to the folder containing the R files.

recursive
If TRUE, subfolders will also be searched for R files.

Value

A table listing the lines per R file containing non-ASCII characters.

[Package OhdsiRTools version 1.10.0 Index]