formatRText {OhdsiRTools}	R Documentation

Format R code

Description

Format R code

Usage

formatRText(text, width.cutoff = 100)

Arguments

text
A character vector with the R code to be formatted.

width.cutoff
Number of characters that each line should be limited to.

Value

A character vector with formatted R code.

Examples

code <- "example <- function(x,foobar){paste(x,foobar)}"
formatted <- formatRText(code)
writeLines(formatted)

[Package OhdsiRTools version 1.10.0 Index]