takeEnvironmentSnapshot {OhdsiRTools}	R Documentation

Take a snapshot of the R environment

Description

Take a snapshot of the R environment

Usage

takeEnvironmentSnapshot(rootPackage)

Arguments

rootPackage
The name of the root package

Details

This function records all versions used in the R environment that are used by one root package. This can be used for example to restore the environment to the state it was when a particular study package was run using the restoreEnvironment function.

Value

A data frame listing all the dependencies of the root package and their version numbers, in the order in which they should be installed.

Examples

snapshot <- takeEnvironmentSnapshot("OhdsiRTools")
snapshot

[Package OhdsiRTools version 2.0.1 Index]