updatePackageNameFolder {OhdsiRTools}	R Documentation

Update the package name in all R and SQL files in a folder

Description

Update the package name in all R and SQL files in a folder

Usage

updatePackageNameFolder(path = ".", packageName, recursive = TRUE)

Arguments

path
Path to the folder containing the files to update. Only files with the .R and .SQL extension will be updated.

packageName
The replacement package name

recursive
Include all subfolders?

[Package OhdsiRTools version 1.10.0 Index]